// src/types/contracts/XvsVault.ts
export interface XvsVault extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): XvsVault;
  clone(): XvsVault;
  methods: {
    DELEGATION_TYPEHASH(): NonPayableTransactionObject<string>;

    DOMAIN_TYPEHASH(): NonPayableTransactionObject<string>;

    _become(xvsVaultProxy: string): NonPayableTransactionObject<void>;

    add(
      _rewardToken: string,
      _allocPoint: number | string | BN,
      _token: string,
      _rewardPerBlock: number | string | BN,
      _lockPeriod: number | string | BN
    ): NonPayableTransactionObject<void>;

    admin(): NonPayableTransactionObject<string>;

    burnAdmin(): NonPayableTransactionObject<void>;

    checkpoints(
      arg0: string,
      arg1: number | string | BN
    ): NonPayableTransactionObject<{
      fromBlock: string;
      votes: string;
      0: string;
      1: string;
    }>;

    delegate(delegatee: string): NonPayableTransactionObject<void>;

    delegateBySig(
      delegatee: string,
      nonce: number | string | BN,
      expiry: number | string | BN,
      v: number | string | BN,
      r: string | number[],
      s: string | number[]
    ): NonPayableTransactionObject<void>;

    delegates(arg0: string): NonPayableTransactionObject<string>;

    deposit(
      _rewardToken: string,
      _pid: number | string | BN,
      _amount: number | string | BN
    ): NonPayableTransactionObject<void>;

    executeWithdrawal(
      _rewardToken: string,
      _pid: number | string | BN
    ): NonPayableTransactionObject<void>;

    getAdmin(): NonPayableTransactionObject<string>;

    getCurrentVotes(account: string): NonPayableTransactionObject<string>;

    getEligibleWithdrawalAmount(
      _rewardToken: string,
      _pid: number | string | BN,
      _user: string
    ): NonPayableTransactionObject<string>;

    getPriorVotes(
      account: string,
      blockNumber: number | string | BN
    ): NonPayableTransactionObject<string>;

    getRequestedAmount(
      _rewardToken: string,
      _pid: number | string | BN,
      _user: string
    ): NonPayableTransactionObject<string>;

    getUserInfo(
      _rewardToken: string,
      _pid: number | string | BN,
      _user: string
    ): NonPayableTransactionObject<{
      amount: string;
      rewardDebt: string;
      pendingWithdrawals: string;
      0: string;
      1: string;
      2: string;
    }>;

    getWithdrawalRequests(
      _rewardToken: string,
      _pid: number | string | BN,
      _user: string
    ): NonPayableTransactionObject<[string, string][]>;

    massUpdatePools(_rewardToken: string): NonPayableTransactionObject<void>;

    nonces(arg0: string): NonPayableTransactionObject<string>;

    numCheckpoints(arg0: string): NonPayableTransactionObject<string>;

    pendingAdmin(): NonPayableTransactionObject<string>;

    pendingReward(
      _rewardToken: string,
      _pid: number | string | BN,
      _user: string
    ): NonPayableTransactionObject<string>;

    pendingXVSVaultImplementation(): NonPayableTransactionObject<string>;

    poolInfos(
      arg0: string,
      arg1: number | string | BN
    ): NonPayableTransactionObject<{
      token: string;
      allocPoint: string;
      lastRewardBlock: string;
      accRewardPerShare: string;
      lockPeriod: string;
      0: string;
      1: string;
      2: string;
      3: string;
      4: string;
    }>;

    poolLength(rewardToken: string): NonPayableTransactionObject<string>;

    requestWithdrawal(
      _rewardToken: string,
      _pid: number | string | BN,
      _amount: number | string | BN
    ): NonPayableTransactionObject<void>;

    rewardTokenAmountsPerBlock(
      arg0: string
    ): NonPayableTransactionObject<string>;

    set(
      _rewardToken: string,
      _pid: number | string | BN,
      _allocPoint: number | string | BN
    ): NonPayableTransactionObject<void>;

    setNewAdmin(newAdmin: string): NonPayableTransactionObject<void>;

    setRewardAmountPerBlock(
      _rewardToken: string,
      _rewardAmount: number | string | BN
    ): NonPayableTransactionObject<void>;

    setWithdrawalLockingPeriod(
      _rewardToken: string,
      _pid: number | string | BN,
      _newPeriod: number | string | BN
    ): NonPayableTransactionObject<void>;

    setXvsStore(
      _xvs: string,
      _xvsStore: string
    ): NonPayableTransactionObject<void>;

    totalAllocPoints(arg0: string): NonPayableTransactionObject<string>;

    updatePool(
      _rewardToken: string,
      _pid: number | string | BN
    ): NonPayableTransactionObject<void>;

    xvsAddress(): NonPayableTransactionObject<string>;

    xvsStore(): NonPayableTransactionObject<string>;

    xvsVaultImplementation(): NonPayableTransactionObject<string>;
  };
// ...
}